ai/cmd/classifier/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
onnxruntime>=1.19.0
pydantic>=2.9.0
grpcio>=1.66.0
grpcio-tools>=1.66.0
